# Conflicts:
#	src/main/java/sarik/dev/foodwaveproject/controller/ProductController.java
#	src/main/java/sarik/dev/foodwaveproject/service/ProductService.java
#	src/main/java/sarik/dev/foodwaveproject/service/impl/ProductServiceImpl.java

+++ src/main/java/sarik/dev/foodwaveproject/entity/Order.java
@@ -42,6 +42,6 @@ public class Order {
     @JoinColumn(name = "payment_id")
     private Payment payment;
 
-    private Double totalAmount;
+    private Long totalAmount;
     private String orderStatus;
 }
